async player stats
rest call to ball-reference.com takes around 2-3s. Actual time to retrieve all the data takes about 25s. Spring creates several threads to process all the boxscore links. Want to add changes to the async support of players to reflect the process of whats happening in statimpl.
for example, even though the call to retrieve all the data is v quick, in the background the threads are asynchronously processing all the results so not everything is available at one. cant get the stats of a player if the code hasnt reached it yet so I should modify playerimpl by adding a completable future to a call to manage that.